+++ ordering/styles/OrderCreator.js
@@ -1,0 +1,5 @@
+export default {
+  labelText: () => ({
+    textTransform: 'capitalize'
+  })
+}
