+++ tasks/getters/getStartOrderTaskProps.js
@@ -1,20 +1,8 @@
 import { createStructuredSelector } from 'reselect'
-import getOrders from '../../ordering/getters/getOrders'
 import getCurrentAgent from 'dogstack-agents/agents/getters/getCurrentAgent'
-import getCurrentAgentGroupIds from '../../agents/getters/getCurrentAgentGroupIds'
-import getCurrentAgentGroupSupplierIds from '../../agents/getters/getCurrentAgentGroupSupplierIds'
-import getCurrentAgentGroupMemberIds from '../../agents/getters/getCurrentAgentGroupMemberIds'
-import getCurrentAgentGroupProfiles from '../../agents/getters/getCurrentAgentGroupProfiles'
-import getCurrentAgentGroupSupplierProfiles from '../../agents/getters/getCurrentAgentGroupSupplierProfiles'
-import getCurrentAgentGroupMemberProfiles from '../../agents/getters/getCurrentAgentGroupMemberProfiles'
+import getCurrentOrder from '../../ordering/getters/getCurrentOrder'
 
 export default createStructuredSelector({
   currentAgent: getCurrentAgent,
-  orders: getOrders,
-  currentAgentGroupIds: getCurrentAgentGroupIds,
-  currentAgentGroupSupplierIds: getCurrentAgentGroupSupplierIds,
-  currentAgentGroupMemberIds: getCurrentAgentGroupMemberIds,
-  currentAgentGroupProfiles: getCurrentAgentGroupProfiles,
-  currentAgentGroupSupplierProfiles: getCurrentAgentGroupSupplierProfiles,
-  currentAgentGroupMemberProfiles: getCurrentAgentGroupMemberProfiles
+  currentOrder: getCurrentOrder
 })
