--- tasks/styles/DashboardTasks.js
@@ -1,18 +1,0 @@
-export default {
-  container: () => ({
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center'
-  }),
-  labelText: () => ({
-    textTransform: 'capitalize'
-  }),
-  intro: ({ theme }) => ({
-    textAlign: 'center',
-    fontSize: theme.fontSizes[2]
-  }),
-  taskNameText: () => ({
-    textTransform: 'capitalize'
-  })
-}
